Move [notein] tests to glue-tests.js
